Fix typos in item-controller
